Handle Velocity kicking us for logging in from a different instance
Handle and send notifications when the proxy is unable to log on because the account is already logged in to the server ("You are already connected to this proxy!"); as opposed to the previous behavior of just silently crashing.

Afaik, this is a Velocity thing that's intentionally enabled on 2b2t to stop exploits.