Added debug-info command
Streamlines the process of submitting bug reports by allowing users to copy-and-paste detailed information about their copy of 2based2wait with minimal information about their system (they can elaborate in their report if they're comfortable).

The script generates valid git sha1 hashes that allow us to figure out what commit hash someone is on even if they didn't clone with git. A basic utility to do that is provided in ./scripts/findCommitBySubstringInFileHash.sh, though it could be modified slightly to only return a single result or even (uselessly) verify file integrity.

Mocha tests are also ran (if mocha is installed), and the config.json is tested for whether it is valid json5.