feat/start: Begin work on high-level logging
This commit adds logging for when a player joins/leaves the server and
enters/leaves the bot's render distance.

The end goal is to create a highly customizable and human-readable logging
category that has widespread coverage of dozens of Mineflayer events.

One problem with this is that balancing readability and information is
really difficult with the current logging solution--but I don't want to
create a novel file format to store every bit of information that I
want.

The solution I think I'll aim for is to favor readability for now and cut
out all the juicy info and *instead* look into satisfying the want to log
everything with the mcpr file format, aka the Minecraft Replay Mod file
format

So far the only source of info on the format I've found is:
https://github.com/ReplayMod/ReplayStudio, which doesn't exactly have a
design doc in it--I have literally no idea if this will go well or not.

One downside of this, though, is that you can't exactly grep through a
mcpr file the same way you would a plaintext 2based2wait log file.