feat: experimental.syncGamestate for desyncs
experimental.syncGamestate is an experimental flag that aggressively
sends fake packets in an attempt to sync the client's game state.

This is to help with clients getting desynced when they miss packets
that only get sent once, either by connecting to the proxy after it has
joined a game OR by logging in and out of the proxy.

Relevant Issue: #85

Current features:
1. Fix invisible vehicles
2. Forcibly update experience bar
3. Fix undismountable vehicles

TODO: (things to fix)
- Sign data desync (blank signs)
- Hunger
- Empty Advancements
- Empty Recipe book (from a cursory glance this appears to already
  be fixed upstream, not sure why this is broken)
- Unre-mountable vehicles after reconnecting
	* Probably something to with metadata or dismounting.
	* Breaking boat and placing it down again fixed this